[refactor] renamed lots of files to follow a better naming convention
