+++ src/core/database/database.types.ts
@@ -1,8 +1,8 @@
 import { JsonDatabase } from "@/database/json-database.js";
-import { AccountRawSchema } from "@/features/accounts/schemas/accounts-schemas.js";
-import { CurrencyRawSchema } from "@/features/currencies/schemas/currencies-schemas.js";
-import { ItemRawSchema } from "@/features/items/schemas/items-schemas.js";
-import { ShopRawSchema } from "@/features/shops/schemas/shop-schemas.js";
+import { AccountRawSchema } from "@/features/accounts/schemas/accounts.schemas.js";
+import { CurrencyRawSchema } from "@/features/currencies/schemas/currencies.schemas.js";
+import { ItemRawSchema } from "@/features/items/schemas/items.schemas.js";
+import { ShopRawSchema } from "@/features/shops/schemas/shop.schemas.js";
 import { NanoIdSchema, SnowflakeSchema } from "@/schemas/utils.js";
 
 export type CurrenciesDatabase = JsonDatabase<typeof NanoIdSchema, typeof CurrencyRawSchema>
