+++ src/core/database/hydrator.ts
@@ -1,9 +1,9 @@
 
-import { NanoId, ApiError, Balance } from "@/database/database-types.js"
-import { Account } from "@/features/accounts/database/accounts-type.js"
-import { Currency } from "@/features/currencies/database/currencies-types.js"
-import { Item } from "@/features/items/database/items-types.js"
-import { Product } from "@/features/shops/database/products-types.js"
+import { NanoId, ApiError, Balance } from "@/database/database.types.js"
+import { Account } from "@/features/accounts/database/accounts.type.js"
+import { Currency } from "@/features/currencies/database/currencies.types.js"
+import { Item } from "@/features/items/database/items.types.js"
+import { Product } from "@/features/shops/database/products.types.js"
 import { assertNeverReached, err, ok } from "@/lib/error-handling.js"
 import { Labelled, Identifiable, Emojiable } from "@/lib/types/core.js"
 import { AwaitedObjectResultReturn } from "@/lib/types/helpers.js"
