+++ src/core/database/init-databases.ts
@@ -1,7 +1,7 @@
-import { AccountRawSchema } from "@/features/accounts/schemas/accounts-schemas.js"
-import { CurrencyRawSchema } from "@/features/currencies/schemas/currencies-schemas.js"
-import { ItemRawSchema } from "@/features/items/schemas/items-schemas.js"
-import { ShopRawSchema } from "@/features/shops/schemas/shop-schemas.js"
+import { AccountRawSchema } from "@/features/accounts/schemas/accounts.schemas.js"
+import { CurrencyRawSchema } from "@/features/currencies/schemas/currencies.schemas.js"
+import { ItemRawSchema } from "@/features/items/schemas/items.schemas.js"
+import { ShopRawSchema } from "@/features/shops/schemas/shop.schemas.js"
 import { NanoIdSchema, SnowflakeSchema } from "@/schemas/utils.js"
 import fs from "fs/promises"
 import { Hydrator } from "./hydrator.js"
