+++ src/features/accounts/database/accounts.type.ts
@@ -1,0 +1,4 @@
+import z from "zod"
+import { AccountRawSchema } from "../schemas/accounts.schemas.js"
+
+export type Account = z.infer<typeof AccountRawSchema>
