+++ src/features/accounts/user-flows/account-give.ts
@@ -1,8 +1,8 @@
 import { t } from "@/core/i18n/i18n.js"
 import { getOrCreateAccount } from "@/core/services/accounts/accounts.services.js"
 import { getCurrencies } from "@/core/services/currencies/currencies.services.js"
-import { NanoId } from "@/database/database-types.js"
-import { Currency } from "@/features/currencies/database/currencies-types.js"
+import { NanoId } from "@/database/database.types.js"
+import { Currency } from "@/features/currencies/database/currencies.types.js"
 import { logToDiscord, replyErrorMessage, updateAsErrorMessage, updateAsSuccessMessage } from "@/lib/discord.js"
 import { Identifiable } from "@/lib/types/core.js"
 import { ExtendedButtonComponent } from "@/lib/ui/ui-components/button.js"
@@ -12,7 +12,7 @@ import { UserFlow } from "@/lib/ui/user-flows/user-flow.js"
 import { validate } from "@/lib/validation.js"
 import { SnowflakeSchema } from "@/schemas/utils.js"
 import { APIRole, ButtonInteraction, ButtonStyle, ChatInputCommandInteraction, MessageFlags, Role, User, bold, roleMention, userMention } from "discord.js"
-import { setAccountCurrencyAmount } from "../services/accounts-services.js"
+import { setAccountCurrencyAmount } from "../services/accounts.services.js"
 
 
 export class AccountGiveFlow extends UserFlow {
