+++ src/features/currencies/database/currencies.types.ts
@@ -1,4 +1,4 @@
 import z from "zod"
-import { CurrencyRawSchema } from "../schemas/currencies-schemas.js"
+import { CurrencyRawSchema } from "../schemas/currencies.schemas.js"
 
 export type Currency = z.infer<typeof CurrencyRawSchema>
