+++ src/features/items/database/items.types.ts
@@ -1,4 +1,4 @@
 import z from "zod";
-import { ItemRawSchema } from "../schemas/items-schemas.js";
+import { ItemRawSchema } from "../schemas/items.schemas.js";
 
 export type Item = z.infer<typeof ItemRawSchema>
