+++ src/features/settings/database/settings.database.ts
@@ -1,9 +1,9 @@
 import fs from "fs/promises"
 import { err, ok } from "@/lib/error-handling.js"
 import { EVENTS } from "@/core/events/event-bus.js"
-import { SettingSchema } from "../schemas/settings-schemas.js"
+import { SettingSchema } from "../schemas/settings.schemas.js"
 import z from "zod"
-import { SettingIdBrands, SettingValueByIdBrand } from "./settings-types.js"
+import { SettingIdBrands, SettingValueByIdBrand } from "./settings.types.js"
 import { JsonDatabase } from "@/database/json-database.js"
 
 const settingsDatabasePath = "./data/settings.json"
