+++ src/features/shops/schemas/shop.schemas.ts
@@ -1,41 +1,12 @@
 import { EmojiSchema, NanoIdSchema, SnowflakeSchema } from "@/schemas/utils.js"
 import z from "zod"
+import { ProductRawSchema } from "./products.schemas.js"
 
 export const SHOP_NAME_MAX_LENGTH = 120
 export const SHOP_DESCRIPTION_MAX_LENGTH = 480
 
 export const DISCOUNT_CODE_MIN_LENGTH = 6
 export const DISCOUNT_CODE_MAX_LENGTH = 8
-
-export const ProductActionSchema = z.discriminatedUnion("type", [
-    z.object({ 
-        type: z.literal("give-role"),
-        options: z.object({
-            roleId: SnowflakeSchema
-        })
-    }),
-    z.object({
-        type: z.literal("give-currency"),
-        options: z.object({
-            currencyId: NanoIdSchema,
-            amount: z.number().min(0)
-        })
-    })
-])
-
-
-export const ProductRawSchema = z.object({
-    itemId: NanoIdSchema,
-    price: z.record(
-        NanoIdSchema, 
-        z.number().min(0)
-    ),
-    stock: z.exactOptional(z.nullable(z.number().min(0))),
-    
-    action: z.exactOptional(z.nullable(ProductActionSchema))
-})
-
-
 
 export const ShopRawSchema = z.object({
     name: z.string()
@@ -63,5 +34,3 @@ export const ShopRawSchema = z.object({
         ProductRawSchema
     )
 })
-
-export type ShopRaw = z.infer<typeof ShopRawSchema>
