+++ src/features/shops/user-flows/shop-create.ts
@@ -1,7 +1,7 @@
 import { t } from "@/core/i18n/i18n.js"
 import { getCurrencies } from "@/core/services/currencies/currencies.services.js"
 import { createShop } from "@/core/services/shops/shops.services.js"
-import { Currency } from "@/features/currencies/database/currencies-types.js"
+import { Currency } from "@/features/currencies/database/currencies.types.js"
 import { replyErrorMessage, updateAsErrorMessage, updateAsSuccessMessage } from "@/lib/discord.js"
 import { ExtendedButtonComponent } from "@/lib/ui/ui-components/button.js"
 import { ExtendedComponent } from "@/lib/ui/ui-components/extended-components.js"
