+++ src/features/shops/user-interfaces/buy.ts
@@ -1,7 +1,7 @@
 import { HYDRATOR } from "@/core/database/init-databases.js"
 import { t } from "@/core/i18n/i18n.js"
 import { processPurchase } from "@/core/services/shops/buy.js"
-import { NanoId } from "@/database/database-types.js"
+import { NanoId } from "@/database/database.types.js"
 import { logToDiscord, replyErrorMessage, updateAsErrorMessage, updateAsSuccessMessage } from "@/lib/discord.js"
 import { assertNeverReached } from "@/lib/error-handling.js"
 import { PrettyLog } from "@/lib/pretty-log.js"
@@ -16,8 +16,8 @@ import { formattedEmojiableName } from "@/utils/formatting.js"
 import { stringifyObj } from "@/utils/objects.js"
 import { bold, ButtonInteraction, ButtonStyle, GuildMember } from "discord.js"
 import z from "zod"
-import { Product } from "../database/products-types.js"
-import { Shop } from "../database/shops-types.js"
+import { Product } from "../database/products.types.js"
+import { Shop } from "../database/shops.types.js"
 import { applyQuantity, formatPrice } from "../services/price.js"
