+++ src/tools/migrate-db-to-v3.ts
@@ -1,8 +1,8 @@
-import { NanoId } from "@/database/database-types.js"
-import { AccountRawSchema } from "@/features/accounts/schemas/accounts-schemas.js"
-import { CurrencyRawSchema } from "@/features/currencies/schemas/currencies-schemas.js"
-import { ItemRawSchema } from "@/features/items/schemas/items-schemas.js"
-import { ProductActionSchema, ProductRawSchema, ShopRawSchema } from "@/features/shops/schemas/shop-schemas.js"
+import { NanoId } from "@/database/database.types.js"
+import { AccountRawSchema } from "@/features/accounts/schemas/accounts.schemas.js"
+import { CurrencyRawSchema } from "@/features/currencies/schemas/currencies.schemas.js"
+import { ItemRawSchema } from "@/features/items/schemas/items.schemas.js"
+import { ProductActionSchema, ProductRawSchema, ShopRawSchema } from "@/features/shops/schemas/shop.schemas.js"
 import { PrettyLog } from "@/lib/pretty-log.js"
 import { validate } from "@/lib/validation.js"
 import fs from "fs/promises"
